Moved `HKD` into `cardano-ledger-core`
It turns out there was two copies of `HKD` defined in
`cardano-ledger-shelley`. Consolidated into one and moved it into core.

It would actually make sense to move this tool into `cardano-base`,
since it also turns out there is another copy of this concept defined
in `ouroboros-consensus/src/Ouroboros/Consensus/Util/Args.hs`, with a
very similar looking mapping functionality as well. If not move it into
`cardano-base` at least reuse one from ledger in consensus would be a
decent idea too.
